pyproject.toml
[tool.poetry]
name = "Scheduler"
version = "0.1.0"
description = ""
authors = ["Jurriaan Den Toonder <1493561+Fastjur@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11 <3.12"
arrow = "^1.2.3"
python-dotenv = "^1.0.0"
requests-cache = "^1.1.0"
django = "^4.2.5"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
openshift = "^0.13.2"
boto3 = "^1.28.40"
botocore = "^1.31.40"
whitenoise = "^6.5.0"
psycopg = {extras = ["binary"], version = "^3.1.10"}
celery = {extras = ["librabbitmq", "redis"], version = "^5.3.4"}
humanize = "^4.8.0"
cx-oracle = "^8.3.0"
docker = "^6.1.3"
entsoe-py = "^0.5.10"
dynaconf = "^3.2.3"
enlighten = "^1.12.2"
django-extensions = "^3.2.3"
scipy = "^1.11.4"

[tool.poetry.group.dev.dependencies]
pylint-runner = "^0.6.0"
black = "^22.10.0"
flake8 = "^3.8.4"
bandit = "^1.7.4"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pytest = "^7.2.0"
pytest-profiling = "^1.7.0"
isort = "^5.10.1"
flake8-isort = "^6.0.0"
flake8-black = "^0.3.6"
matplotlib = "^3.7.2"
tqdm = "^4.66.1"
pillow = "^10.0.0"
pydevd-pycharm = ">=232.9559.58,<232.9560.0"
pytest-django = "^4.5.2"
seaborn = {extras = ["stats"], version = "^0.13.0"}
notebook = "^7.0.6"
django-extensions = "^3.2.3"
statsmodels = "^0.14.0"


[tool.poetry.group.synthetics]
optional = true

[tool.poetry.group.synthetics.dependencies]
py7zr = "^0.20.5"
pandas = "^2.0.2"
tqdm = "^4.65.0"
django = "^4.2.5"
requests = "^2.31.0"
botocore = "^1.31.45"
boto3 = "^1.28.45"
matplotlib = "^3.7.3"
pydevd-pycharm = ">=232.9559.58,<232.9560.0"
nltk = "^3.8.1"
gensim = "^4.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
